features acheived
-list configuration and choose configuration
-only load connector that are mapped to configuration
-each user has their own environment on login
