+++ app/api/v1/connector.py
@@ -27,7 +27,7 @@ inference_router = APIRouter()
 actions = APIRouter()
 
 @router.get("/list", response_model=resp_schemas.CommonResponse, dependencies=[Depends(verify_token)])
-def list_connectors(db: Session = Depends(get_db), provider_category_ids:  Optional[List[int]] = None ):
+def list_connectors(db: Session = Depends(get_db), provider_category_ids:  Optional[List[int]] = None, user_data: dict = Depends(verify_token)):
 
     """
     Retrieves a list of all connectors from the database. If a provider category ID is provided, only connectors from that category are returned.
@@ -38,11 +38,11 @@ def list_connectors(db: Session = Depends(get_db), provider_category_ids:  Optio
     Returns:
         CommonResponse: A response containing either the list of connectors or an error message.
     """
-
+    user_id = user_data["user_id"]
     if provider_category_ids:
-        result, error = svc.list_connectors_by_provider_category(provider_category_ids, db)
+        result, error = svc.list_connectors_by_provider_category(provider_category_ids, db, user_id)
     else:
-        result, error = svc.list_connectors(db)
+        result, error = svc.list_connectors(db, user_id)
 
     if error:
         return commons.is_error_response("DB Error", result, {"connectors": []})
@@ -122,7 +122,7 @@ async def upload_document_datsource(
     )
 
 @router.post("/create", response_model=resp_schemas.CommonResponse, dependencies=[Depends(verify_token)])
-def create_connector(connector: schemas.ConnectorBase, db: Session = Depends(get_db)):
+def create_connector(connector: schemas.ConnectorBase, db: Session = Depends(get_db), user_data: dict = Depends(verify_token)):
 
     """
     Creates a new connector in the database.
@@ -134,8 +134,8 @@ def create_connector(connector: schemas.ConnectorBase, db: Session = Depends(get
     Returns:
         CommonResponse: A response indicating success or failure of the connector creation process.
     """
-
-    result, error = svc.create_connector(connector, db)
+    user_id = user_data["user_id"]
+    result, error = svc.create_connector(connector, db, user_id)
     if error:
         return commons.is_error_response("Connector Not Created", error, {"connector": {}})
 
@@ -242,7 +242,7 @@ def updateschemas(connector_id: int, connector: schemas.SchemaUpdate, db: Sessio
 
 
 @router.get("/configuration/list", response_model=resp_schemas.CommonResponse, dependencies=[Depends(verify_token)])
-def list_configurations(db: Session = Depends(get_db)):
+def list_configurations(db: Session = Depends(get_db), user_data: dict = Depends(verify_token)):
 
     """
     Lists all available configurations from the database.
@@ -253,8 +253,8 @@ def list_configurations(db: Session = Depends(get_db)):
     Returns:
         CommonResponse: A response containing the list of configurations or an error message.
     """
-
-    result, error = svc.list_configurations(db)
+    user_id = user_data["user_id"]
+    result, error = svc.list_configurations(db, user_id)
 
     if error:
         return commons.is_error_response("DB error", result, {"configurations": []})
@@ -269,9 +269,66 @@ def list_configurations(db: Session = Depends(get_db)):
         error=None,
         data={"configurations": result}
     )
+    
+@router.get("/configuration/{config_id}", response_model=resp_schemas.CommonResponse, dependencies=[Depends(verify_token)])
+def get_configuration(config_id: int, db: Session = Depends(get_db)):
+    """
+    Retrieves a configuration by its ID.
+
+    Args:
+        config_id (int): ID of the configuration to retrieve.
+        db (Session): Database session dependency.
+
+    Returns:
+        CommonResponse: A response containing the configuration or an error message.
+    """
+    result, error = svc.get_configuration(db, config_id)
+
+    if error == "DB Error":
+        return commons.is_error_response("DB error", result, {"configuration": None})
+
+    if error == "Configuration not found":
+        return commons.is_none_reponse("Configuration Not Found", {"configuration": None})
+
+    return resp_schemas.CommonResponse(
+        status=True,
+        status_code=200,
+        message="Configuration retrieved successfully",
+        error=None,
+        data={"configuration": result}
+    )
+    
+@router.delete("/configuration/{config_id}", response_model=resp_schemas.CommonResponse, dependencies=[Depends(verify_token)])
+def get_configuration(config_id: int, db: Session = Depends(get_db)):
+    """
+    Retrieves a configuration by its ID.
+
+    Args:
+        config_id (int): ID of the configuration to retrieve.
+        db (Session): Database session dependency.
+
+    Returns:
+        CommonResponse: A response containing the configuration or an error message.
+    """
+    result, error = svc.delete_configuration(db, config_id)
+
+    if error == "DB Error":
+        return commons.is_error_response("DB error", result, {"configuration": None})
+
+    if error == "Configuration not found":
+        return commons.is_none_reponse("Configuration Not Found", {"configuration": None})
+
+    return resp_schemas.CommonResponse(
+        status=True,
+        status_code=200,
+        message="Configuration deleted successfully",
+        error=None,
+        data={"configuration": result}
+    )
+
 
 @router.post("/configuration/create", response_model=resp_schemas.CommonResponse, dependencies=[Depends(verify_token)])
-def create_configuration(configuration: schemas.ConfigurationCreation, db: Session = Depends(get_db)):
+def create_configuration(configuration: schemas.ConfigurationCreation, db: Session = Depends(get_db), user_data: dict = Depends(verify_token)):
 
     """
     Creates a new configuration and stores it in the database.
@@ -283,8 +340,8 @@ def create_configuration(configuration: schemas.ConfigurationCreation, db: Sessi
     Returns:
         CommonResponse: A response indicating the success or failure of the configuration creation.
     """
-
-    result, error = svc.create_configuration(configuration, db)
+    user_id = user_data["user_id"]
+    result, error = svc.create_configuration(configuration, db, user_id)
 
     if error:
         return commons.is_error_response("DB error", result, {"configuration": []})
@@ -517,7 +574,7 @@ def create_yaml(request: Request, config_id: int, db: Session = Depends(get_db))
     config["datasources"] = data_sources
     config["models"] = inference_config
 
-    confyaml = svc.get_inference_and_plugin_configurations(db)
+    confyaml = svc.get_inference_and_plugin_configurations(db, config_id)
     request.app.container.config.from_dict(confyaml)
     datasources = request.app.container.datasources()
