+++ ui/src/layouts/dashboard/DashboadBody.jsx
@@ -1,15 +1,35 @@
 
 import help from "src/assets/icons/help.svg"
 import style from "./Dashboard.module.css"
+import { restartBot } from "src/services/BotConfifuration";
+import { toast } from "react-toastify";
 
-const DashboardBody = ({ title = "Dashboard",children, containerStyle = {}, containerClassName = ""}) => {
+const DashboardBody = ({ title = "Dashboard",options=[], select, children, containerStyle = {}, containerClassName = ""}) => {
+    const handleRestart = (event) => {
+        const selectedId = event.target.value;
+        restartBot(selectedId)
+            .then(() => {
+                toast.success(`Configuration ${selectedId} loaded`);
+            })
+            .catch(() => {
+                toast.error("Failed to load configuration");
+            });
+    };
     return (
         <>
             <div className={`${style.DashboardBody} ${containerClassName}`} >
                 
                 <div className={style.DashboardHeader}>
                     <div className={style.DashboardTitleContainer}>
-                        <span className={style.DashboardTitle}>{title}</span>
+                        {select ? (
+                            <select onChange={handleRestart} defaultValue="">
+                                <option value="" disabled>Select a Configuration</option>
+                                {options.map((id) => (
+                                    <option key={id} value={id}>Configuration {id}</option>
+                                ))}
+                            </select>
+                        ) : ( 
+                        <span className={style.DashboardTitle}>{title}</span> )}
                     </div>
                     <div>
                     </div>
