+++ ui/src/pages/ChatConfiguration/Configuration.module.css
@@ -1,0 +1,39 @@
+/* Empty Configuration Page */
+
+.EmptyDataContainer {
+    text-align: center;
+    margin-top: 20vh;
+}
+
+.EmptyDataTitleSpan {
+    color: var(--neural-3);
+    text-align: center;
+    font-feature-settings: 'clig' off, 'liga' off;
+    font-family: Inter;
+    font-size: 16px;
+    font-style: normal;
+    font-weight: 400;
+    line-height: 20px;
+}
+
+/* end Empty Configuration Page ****/
+
+
+/* Config List Page */
+
+.SearchContainer {
+    margin-bottom: 44px;
+    display: flex;
+}
+
+.ConnectorAction {
+    display: flex;
+    align-items: center;
+    gap: 23px;
+}
+
+.ConnectorIcon {
+    width: 24px;
+    height: 24px;
+    margin-right: 8px;
+}
